refactor: mv useNativeCurrency to lib/hooks
