+++ src/lib/hooks/useNativeCurrency.ts
@@ -1,0 +1,17 @@
+import { Currency } from '@uniswap/sdk-core'
+import { SupportedChainId } from 'constants/chains'
+import { nativeOnChain } from 'constants/tokens'
+import useActiveWeb3React from 'hooks/useActiveWeb3React'
+import { useMemo } from 'react'
+
+export default function useNativeCurrency(): Currency {
+  const { chainId } = useActiveWeb3React()
+  return useMemo(
+    () =>
+      chainId
+        ? nativeOnChain(chainId)
+        : // display mainnet when not connected
+          nativeOnChain(SupportedChainId.MAINNET),
+    [chainId]
+  )
+}
